Better code quality, documentation and README
