+++ .vscode/settings.json
@@ -11,77 +11,18 @@
     "cjson.h": "c",
     "format": "c",
     "include.h": "c",
-<<<<<<< Updated upstream
     "string.h": "c",
     "stdio.h": "c",
     "mcpwm_prelude.h": "c",
     "mcpwm_timer.h": "c",
     "spi_master.h": "c",
-    "pulse_cnt.h": "c"
-=======
-    "adc.h": "c",
-    "adc_types_legacy.h": "c",
-    "gpio.h": "c",
-    "esp_rom_gpio.h": "c",
-    "gpio_etm.h": "c",
-    "adc_types.h": "c",
-    "i2s_hal.h": "c",
-    "i2s_types_legacy.h": "c",
-    "clk_tree_ll.h": "c",
-    "adc_i2s_legacy.h": "c",
-    "adc1_private.h": "c",
-    "string.h": "c",
-    "stdbool.h": "c",
-    "esp_types.h": "c",
-    "freertos.h": "c",
-    "queue.h": "c",
-    "sdkconfig.h": "c",
-    "lldesc.h": "c",
-    "gpio_hal.h": "c",
-    "dac_ll.h": "c",
-    "dac_types.h": "c",
-    "adc_share_hw_ctrl.h": "c",
-    "sar_periph_ctrl.h": "c",
-    "soc_caps.h": "c",
-    "esp_err.h": "c",
-    "esp_clock_output.h": "c",
-    "gdma.h": "c",
-    "esp_intr_alloc.h": "c",
-    "clk_ctrl_os.h": "c",
-    "esp_check.h": "c",
-    "esp_attr.h": "c",
-    "esp_pm.h": "c",
-    "esp_efuse.h": "c",
-    "periph_ctrl.h": "c",
-    "esp_efuse_chip.h": "c",
-    "esp_cache.h": "c",
-    "i2s_platform.h": "c",
-    "periph_defs.h": "c",
-    "stdint.h": "c",
-    "esp_clk.h": "c",
-    "ctype.h": "c",
-    "timers.h": "c",
-    "rtc.h": "c",
-    "stdlib.h": "c",
-    "gpio_types.h": "c",
-    "clk_tree_defs.h": "c",
-    "rtc_io.h": "c",
-    "adc_ll.h": "c",
-    "rtc_io_struct.h": "c",
-    "adc_periph.h": "c",
-    "syscon_struct.h": "c",
-    "rtc_cntl_struct.h": "c",
-    "lock.h": "c",
-    "adc_hal_common.h": "c",
-    "adc_types_private.h": "c",
-    "syscon_reg.h": "c",
-    "i2s.h": "c"
->>>>>>> Stashed changes
+    "pulse_cnt.h": "c",
+    "cstdlib": "c"
   },
   "idf.flashType": "UART",
   "idf.port": "/dev/ttyUSB0",
   "idf.adapterTargetName": "esp32",
-  "idf.espIdfPath": "/home/joaquin/esp/v5.4/esp-idf",
-  "idf.toolsPath": "/home/joaquin/.espressif",
+  "idf.espIdfPath": "/home/jotalora/esp/v5.3.1/esp-idf",
+  "idf.toolsPath": "/home/jotalora/.espressif",
   "idf.portWin": "COM3"
 }
