+++ main/CMakeLists.txt
@@ -1,2 +1,4 @@
-idf_component_register(SRCS "main.c"
-                    INCLUDE_DIRS "." "../include")
+idf_component_register(
+    SRCS "main.c" "network.c" "crypto.c" "acquisition.c" "webservers.c" "data_transmission.c"
+    INCLUDE_DIRS "." "../include"
+)
