+++ sdkconfig
@@ -1388,14 +1388,11 @@ CONFIG_HAL_SPI_SLAVE_FUNC_IN_IRAM=y
 CONFIG_HEAP_POISONING_DISABLED=y
 # CONFIG_HEAP_POISONING_LIGHT is not set
 # CONFIG_HEAP_POISONING_COMPREHENSIVE is not set
-# CONFIG_HEAP_TRACING_OFF is not set
-CONFIG_HEAP_TRACING_STANDALONE=y
+CONFIG_HEAP_TRACING_OFF=y
+# CONFIG_HEAP_TRACING_STANDALONE is not set
 # CONFIG_HEAP_TRACING_TOHOST is not set
-CONFIG_HEAP_TRACING=y
-CONFIG_HEAP_TRACING_STACK_DEPTH=2
 # CONFIG_HEAP_USE_HOOKS is not set
 # CONFIG_HEAP_TASK_TRACKING is not set
-# CONFIG_HEAP_TRACE_HASH_MAP is not set
 # CONFIG_HEAP_ABORT_WHEN_ALLOCATION_FAILS is not set
 # CONFIG_HEAP_PLACE_FUNCTION_INTO_FLASH is not set
 # end of Heap memory debugging
